+++ Guanima.Redis/Commands/Hashes/HIncrByCommand.cs
@@ -5,7 +5,7 @@ namespace Guanima.Redis.Commands.Hashes
     [Serializable]
     public class HIncrByCommand : KeyCommand
     {
-        public HIncrByCommand(string key, string field, int value)
+        public HIncrByCommand(string key, string field, long value)
             : base(key, field, value)
         {
             // todo: validate field
